// swipecore/backend/src/types/otpTypes.ts
export interface OtpResponse {
    success: boolean;
    message: string;
    errorCode?: string;
    accessToken?: string; // generated by the server every session
    refreshToken?: string; // generated by the server every OTP login
};

export interface OtpRequest {
    phoneNumber: string;
};

export interface OtpVerifyRequest {
    phoneNumber: string;
    code: string;
};

export interface OtpRefreshRequest {
    refreshToken: string;
};